refactor: combine VerifyProviderListener and VerifyProviderIsMilestoneAwareListener
No need for two listeners.
