+++ src/Milestone/VerifyProviderListener.php
@@ -10,6 +10,8 @@ declare(strict_types=1);
 
 namespace Phly\KeepAChangelog\Milestone;
 
+use Phly\KeepAChangelog\Provider\MilestoneAwareProviderInterface;
+
 class VerifyProviderListener
 {
     public function __invoke(AbstractMilestoneProviderEvent $event) : void
@@ -23,6 +25,11 @@ class VerifyProviderListener
         }
 
         $provider = $providerSpec->createProvider();
+
+        if (! $provider instanceof MilestoneAwareProviderInterface) {
+            $event->providerIncapableOfMilestones();
+            return;
+        }
 
         $event->discoveredProvider($provider);
     }
